feat: src/modules, src/hash
