+++ src/modules/test.js
@@ -1,0 +1,3 @@
+import { unknownObject, createMyServer } from './cjsToEsm.mjs'
+
+console.log(unknownObject, createMyServer)
